Move test-only commands to testing folder
Moved SpinHoldCommand and SpinLaneCommand to testing/ subdirectory.
These are simple debug commands not used in production:
- SpinHoldCommand: Manual spin while button held
- SpinLaneCommand: References deleted FireLaneCommand

All commands actively used in LauncherCommands factory remain in main folder.
