+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/commands/LauncherCommands/testing/SpinHoldCommand.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.commands.LauncherCommands;
+package org.firstinspires.ftc.teamcode.commands.LauncherCommands.testing;
 
 import dev.nextftc.core.commands.Command;
 
@@ -7,15 +7,13 @@ import org.firstinspires.ftc.teamcode.subsystems.LauncherSubsystem;
 import java.util.Objects;
 
 /**
- * Keeps the launcher spinning at full power while the driver holds a lane button.
- * The command never completes on its own; the release logic handles firing via
- * a paired {@link FireLaneCommand}.
+ * Keeps the launcher spinning at full power while a button is held.
  */
-public class SpinLaneCommand extends Command {
+public class SpinHoldCommand extends Command {
 
     private final LauncherSubsystem launcher;
 
-    public SpinLaneCommand(LauncherSubsystem launcher) {
+    public SpinHoldCommand(LauncherSubsystem launcher) {
         this.launcher = Objects.requireNonNull(launcher, "launcher required");
         requires(launcher);
         setInterruptible(true);
@@ -28,7 +26,7 @@ public class SpinLaneCommand extends Command {
 
     @Override
     public void update() {
-        // Nothing to do while the button is held; we just stay in spin mode.
+        // Stay in spin mode while held.
     }
 
     @Override
@@ -38,6 +36,6 @@ public class SpinLaneCommand extends Command {
 
     @Override
     public void stop(boolean interrupted) {
-        // Spin mode will be managed by next command or default to idle
+        launcher.setSpinMode(LauncherSubsystem.SpinMode.IDLE);
     }
 }
