+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/commands/LauncherCommands/testing/SpinLaneCommand.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.commands.LauncherCommands;
+package org.firstinspires.ftc.teamcode.commands.LauncherCommands.testing;
 
 import dev.nextftc.core.commands.Command;
 
@@ -7,13 +7,15 @@ import org.firstinspires.ftc.teamcode.subsystems.LauncherSubsystem;
 import java.util.Objects;
 
 /**
- * Keeps the launcher spinning at full power while a button is held.
+ * Keeps the launcher spinning at full power while the driver holds a lane button.
+ * The command never completes on its own; the release logic handles firing via
+ * a paired {@link FireLaneCommand}.
  */
-public class SpinHoldCommand extends Command {
+public class SpinLaneCommand extends Command {
 
     private final LauncherSubsystem launcher;
 
-    public SpinHoldCommand(LauncherSubsystem launcher) {
+    public SpinLaneCommand(LauncherSubsystem launcher) {
         this.launcher = Objects.requireNonNull(launcher, "launcher required");
         requires(launcher);
         setInterruptible(true);
@@ -26,7 +28,7 @@ public class SpinHoldCommand extends Command {
 
     @Override
     public void update() {
-        // Stay in spin mode while held.
+        // Nothing to do while the button is held; we just stay in spin mode.
     }
 
     @Override
@@ -36,6 +38,6 @@ public class SpinHoldCommand extends Command {
 
     @Override
     public void stop(boolean interrupted) {
-        launcher.setSpinMode(LauncherSubsystem.SpinMode.IDLE);
+        // Spin mode will be managed by next command or default to idle
     }
 }
